+++ Changelog.md
@@ -1,4 +1,4 @@
-# gov-ua-ui changelog from v0.4.5. to v0.4.6.
+# gov-ua-ui changelog from v0.4.6. to v0.4.7.
 
-1. Added className prop to Tooltip component
+1. Added posibility to add components between pagination items and pagination controls in Pagination component.